# Conflicts:
#	CMakeLists.txt
#	Changes.txt
#	Common/ac/gamesetupstruct.cpp
#	Common/core/def_version.h
#	Common/game/interactions.h
#	Common/game/room_file.cpp
#	Common/gui/guibutton.cpp
#	Common/gui/guimain.cpp
#	Common/gui/guimain.h
#	Editor/AGS.Editor/Components/FontsComponent.cs
#	Editor/AGS.Editor/Components/RoomsComponent.cs
#	Editor/AGS.Editor/Resources/agsdefns.sh
#	Editor/AGS.Editor/app.manifest
#	Editor/AGS.Types/Properties/AssemblyInfo.cs
#	Editor/Native/acgui_agsnative.cpp
#	Engine/ac/button.cpp
#	Engine/ac/character.cpp
#	Engine/ac/characterinfo_engine.cpp
#	Engine/ac/dialog.cpp
#	Engine/ac/display.cpp
#	Engine/ac/draw.cpp
#	Engine/ac/dynamicsprite.cpp
#	Engine/ac/dynobj/dynobj_manager.cpp
#	Engine/ac/dynobj/dynobj_manager.h
#	Engine/ac/dynobj/managedobjectpool.cpp
#	Engine/ac/dynobj/managedobjectpool.h
#	Engine/ac/file.cpp
#	Engine/ac/game.cpp
#	Engine/ac/gamestate.cpp
#	Engine/ac/gamestate.h
#	Engine/ac/global_api.cpp
#	Engine/ac/global_audio.cpp
#	Engine/ac/global_character.cpp
#	Engine/ac/global_character.h
#	Engine/ac/global_debug.cpp
#	Engine/ac/global_dialog.cpp
#	Engine/ac/global_dialog.h
#	Engine/ac/global_game.cpp
#	Engine/ac/global_object.cpp
#	Engine/ac/global_overlay.cpp
#	Engine/ac/global_overlay.h
#	Engine/ac/global_video.cpp
#	Engine/ac/global_viewframe.cpp
#	Engine/ac/mouse.cpp
#	Engine/ac/mouse.h
#	Engine/ac/object.cpp
#	Engine/ac/object.h
#	Engine/ac/overlay.cpp
#	Engine/ac/overlay.h
#	Engine/ac/room.cpp
#	Engine/ac/screenoverlay.cpp
#	Engine/ac/screenoverlay.h
#	Engine/game/game_init.cpp
#	Engine/game/savegame.cpp
#	Engine/game/savegame_components.cpp
#	Engine/main/game_run.cpp
#	Engine/main/update.cpp
#	Engine/media/video/video.cpp
#	Engine/script/cc_instance.h
#	Engine/script/script.cpp
#	Engine/script/script.h
#	Engine/script/scriptexecutor.cpp
#	Solutions/Engine.App/Engine.App.vcxproj.filters
#	version.json

+++ Common/game/room_file.cpp
@@ -292,7 +292,7 @@ HError ReadAnimBgBlock(RoomStruct *room, Stream *in, RoomFileVersion data_ver)
     if (room->BgFrameCount > MAX_ROOM_BGFRAMES)
         return new RoomFileError(kRoomFileErr_IncompatibleEngine, String::FromFormat("Too many room backgrounds (in room: %d, max: %d).", room->BgFrameCount, MAX_ROOM_BGFRAMES));
 
-    room->BgAnimSpeed = in->ReadInt8();
+    room->BgAnimSpeed = static_cast<uint8_t>(in->ReadInt8());
     for (size_t i = 0; i < room->BgFrameCount; ++i)
         room->BgFrames[i].IsPaletteShared = in->ReadInt8() != 0;
